+++ IncomeExpense_JavaSpring/src/main/java/TeamHaLoi/IncomeExpenseTracker/repository/BankAccountRepository.java
@@ -5,6 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BankAccountRepository extends JpaRepository<BankAccount, Long> {
+public interface BankAccountRepository extends JpaRepository<BankAccount, Integer> {
     BankAccount findByAccountNumber(String accountNumber);
+
 }
