+++ IncomeExpense_JavaSpring/src/main/java/TeamHaLoi/IncomeExpenseTracker/repository/UserAccountRepository.java
@@ -5,6 +5,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserAccountRepository extends JpaRepository<UserAccount, Long> {
+public interface UserAccountRepository extends JpaRepository<UserAccount, Integer> {
     UserAccount findByEmail(String email);
+
 }
