+++ IncomeExpense_JavaSpring/src/main/java/TeamHaLoi/exception/BankAccountNotFoundException.java
@@ -5,7 +5,7 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(HttpStatus.NOT_FOUND)
 public class BankAccountNotFoundException extends RuntimeException {
-    public BankAccountNotFoundException(Long id) {
+    public BankAccountNotFoundException(Integer id) {
         super("Could not find BankAccount with id: " + id);
     }
 }
