+++ IncomeExpense_JavaSpring/src/main/java/TeamHaLoi/exception/UserAccountNotFoundException.java
@@ -5,7 +5,7 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(HttpStatus.NOT_FOUND)
 public class UserAccountNotFoundException extends RuntimeException {
-    public UserAccountNotFoundException(Long id) {
+    public UserAccountNotFoundException(Integer id) {
         super("Could not find UserAccount with id: " + id);
     }
 }
